+++ private/__main__.py
@@ -2,20 +2,20 @@
 
 def main():
   import sys
-  from modules import helpers
   argv = sys.argv[1:]
   
   if not argv:
-    helpers.show_usage()
+    show_usage()
     
   else:
     from os import path
     from modules import core
+    from modules.helpers import load_module
     
     target_path = argv.pop(0)
     
     if path.isfile(target_path):
-      core.start(helpers.load_module(target_path), argv)
+      core.start(load_module(target_path), argv)
       
     elif path.isdir(target_path):
       from glob import glob
@@ -26,15 +26,18 @@ def main():
           for Module in Modules:
             setattr(self, Module.__name__, Module)
         
-      Container = Container([helpers.load_module(name) for name in glob('%s/*.py' % path.abspath(target_path)) if path.isfile(name)])
+      Container = Container([load_module(name) for name in glob('%s/*.py' % path.abspath(target_path)) if path.isfile(name)])
       
       Group(Container) # brand the Container
       
       core.start(Container, argv)
       
     else:
-      helpers.show_usage()
+      show_usage()
   
+
+def show_usage():
+  print 'Usage:\n\t$ private script/dir [flags] [command] [args]'
 
 if __name__ == '__main__':
   main()