+++ private/types/regex.py
@@ -7,10 +7,9 @@ class pattern(CustomType):
   Used to restrict the input to specific patterns.
   """
   def __init__(self, pattern, flags=0, desc=None):
+    CustomType.__init__(self, desc)
     self.exp = re.compile(pattern, flags)
-    if desc is not None:
-      self.desc = desc
-      
+    
   def __call__(self, val):
     if not self.exp.match(val):
       raise ValueError('Invalid value.')
@@ -24,3 +23,4 @@ email = pattern('^([a-z0-9_\.-]+)@([\da-z\.-]+)\.([a-z\.]{2,6})$', desc='Email')
 username = pattern('^[a-z0-9_-]{3,16}$', desc='Username')
 password = pattern('^[A-z0-9_-]{6,18}$', desc='Password')
 slug = pattern('^[a-z0-9-]+$', desc='Slug')
+number = pattern('^[0-9]+$', desc='Number') # a positive integer